+++ packages/jaspr/lib/src/components/html/html.dart
@@ -1,3 +1,5 @@
+import 'package:universal_web/web.dart' as web;
+
 import '../../foundation/basic_types.dart';
 import '../../foundation/events.dart';
 import '../../foundation/styles/styles.dart';