+++ packages/jaspr/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Added `onClick` override to `a` html tag. Fixing issues with default action handling. @slightfoot
+
 ## 0.18.0
 
 - **BREAKING** Changed `AppBinding`s `Uri get currentUri` to `String get currentUrl`.